[#25] Implement socket in chats & chat window page
